refactor: DispatcherServlet 클래스를 MVC 모듈로 이동 및 ManualHandlerMapping 제거
- DispatcherServlet 및 관련 클래스 MVC 모듈로 리팩토링 및 이동
- ManualHandlerMapping 및 관련 로직 삭제
- AnnotationHandlerMapping 기반 매핑으로 통합
- IndexController 추가 및 루트 경로 핸들링 로직 설정
